fix overflow

+++ include/cache_type.h
@@ -66,8 +66,8 @@ typedef struct {
 typedef struct {
     char* bdev_name;
     unsigned block_size;              // 通常是 512 Bytes
-    unsigned device_size;             // LBA的數量
-    unsigned cache_block_num;         // cache block的數量 (cache block 32KB)
+    unsigned long device_size;        // LBA的數量
+    unsigned long cache_block_num;    // cache block的數量 (cache block 32KB)
     unsigned blocks_per_page;         // 4KB / 512 = 8
     unsigned blocks_per_cache_block;  // 32KB / 512 = 64
 } device;
